requirements.txt
# Core dependencies
rouge-score>=0.1.2
bert-score>=0.3.13
bleurt @ git+https://github.com/google-research/bleurt.git
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
nltk>=3.8
scipy>=1.10.0

# Notebook and visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0
